wasteClassification/requirements.txt
fastapi
python-consul
tensorflow
keras
uvicorn
python-multipart